Refactoring. Use card & carousel where applicable.
