+++ src/app/components/data-view/data-card/data-card.component.html
@@ -1,3 +1,4 @@
+<app-toast></app-toast>
 <div class="text-center" *ngIf="!resultSet || !resultSet.data || resultSet.data.length == 0">
   <span class="text-danger">(no data to show)</span>
 </div>
@@ -5,7 +6,7 @@
 <div class="card-wrapper" *ngIf="resultSet">
 
   <div class="card" *ngFor="let d1 of resultSet.data; let i = index">
-    <!--<a [routerLink]="[]" class="text-danger delete-card btn btn-outline-secondary btn-sm" (dblclick)="deleteRow(d1, i)"><i class="cui-trash"></i></a>-->
+    <a [routerLink]="[]" class="delete-card btn btn-outline-secondary btn-sm" (click)="deleteRow(d1, i)" (mouseleave)="confirm = -1"><i [ngClass]="{ 'cui-trash': i!==confirm, 'fa fa-warning': i===confirm }"></i></a>
     <ul class="list-group list-group-flush">
       <!-- load MM types first -->
       <ng-container *ngFor="let d2 of d1; let j = index">
@@ -20,7 +21,7 @@
             Your browser does not support the audio element.
           </audio>
           <ng-container *ngSwitchCase="'FILE'">
-            <span class="key">{{resultSet.header[j].name}}</span>: <a class="download" [href]="getFileLink(d2)" download><span class="cui-cloud-download"></span></a>
+            <span class="key">{{resultSet.header[j].name}}:</span> <a class="download val" [href]="getFileLink(d2)" download><span class="cui-cloud-download"></span></a>
           </ng-container>
         </li>
       </ng-container>
@@ -29,7 +30,7 @@
       <ng-container *ngFor="let d2 of d1; let j = index">
         <li class="list-group-item" *ngIf="d2 && !_types.isMultimedia(resultSet.header[j].dataType)" [ngSwitch]="resultSet.header[j].dataType">
           <ng-container *ngSwitchDefault>
-            <span class="key">{{resultSet.header[j].name}}</span><span class="val">: {{d2}}</span>
+            <span class="key">{{resultSet.header[j].name}}:</span><span class="val"> {{d2}}</span>
           </ng-container>
         </li>
       </ng-container>
