+++ src/app/components/data-view/data-carousel/data-carousel.component.scss
