+++ src/app/components/data-view/data-table/table-config.ts
