+++ src/app/components/data-view/data-view.component.scss
@@ -1,0 +1,7 @@
+#mode_wrapper{
+  display: flex;
+  justify-content: center;
+  button{
+    margin: 0 0.125rem;
+  }
+}
