+++ src/app/components/data-view/input/input.component.scss
