+++ src/app/components/data-view/models/result-set.model.ts
