+++ src/app/views/querying/explore-by-example/explore-by-example.component.ts
@@ -3,9 +3,9 @@ import * as $ from 'jquery';
 import {EditTableRequest, QueryExplorationRequest, SchemaRequest} from '../../../models/ui-request.model';
 import {CrudService} from '../../../services/crud.service';
 import {LeftSidebarService} from '../../../components/left-sidebar/left-sidebar.service';
-import {ExplorColSet, ExploreSet, ResultSet} from '../../../components/data-table/models/result-set.model';
+import {ExplorColSet, ExploreSet, ResultSet} from '../../../components/data-view/models/result-set.model';
 import {ToastService} from '../../../components/toast/toast.service';
-import {DataTableComponent} from '../../../components/data-table/data-table.component';
+import {DataTableComponent} from '../../../components/data-view/data-table/data-table.component';
 import {SidebarNode} from '../../../models/sidebar-node.model';
 import {ForeignKey, Uml} from '../../uml/uml.model';
 import {  BsModalService, BsModalRef  } from 'ngx-bootstrap/modal';
