+++ src/app/views/querying/graphical-querying/graphical-querying.component.html
@@ -55,6 +55,6 @@
 <div class="row" *ngIf="resultSet">
   <div class="col-lg-12" style="padding-bottom: 1em;">
     <h5>Query Result</h5>
-    <app-data-table [resultSet]="resultSet" [config]="{create:false, update:false, delete:false, sort:false, search: false, exploring: false}" [loading]="loading"></app-data-table>
+    <app-data-view [resultSet]="resultSet" [config]="{create:false, update:false, delete:false, sort:false, search: false, exploring: false}" [loading]="loading"></app-data-view>
   </div>
 </div>
