+++ src/app/views/querying/sql-console/sql-console.component.html
@@ -70,7 +70,7 @@
 
 
                     <div class="col-lg-12 result-wrapper" *ngIf="resultSet && resultSet.data">
-                        <app-data-table [resultSet]="resultSet" [config]="tableConfig"></app-data-table>
+                        <app-data-view [resultSet]="resultSet" [config]="tableConfig"></app-data-view>
                       <p class="text-muted text-center small mt-1 mb-0" *ngIf="resultSet.hasMoreRows">Only {{resultSet.data.length}} rows are being displayed. To fetch more rows, please specify a LIMIT clause.</p>
                     </div>
                 </div>
