+++ src/app/views/table-view/table-view.component.scss
@@ -4,11 +4,3 @@
   right: 2em;
   border-radius: 100%;
 }
-
-#mode_wrapper{
-  display: flex;
-  justify-content: center;
-  button{
-    margin: 0 0.125rem;
-  }
-}
